added styling updates and lazy loading
