+++ src/frontend/components/filters/filters.css
@@ -20,7 +20,7 @@
   border: none;
   outline: none;
   cursor: pointer;
-  transition: all 300ms ease-in-out;
+  transition: all 200ms ease-in;
 }
 .filter_btn:hover {
   background-color: var(--tertiary);
@@ -66,4 +66,14 @@
 }
 .filter__option__input {
   display: none;
+}
+
+@media screen and (min-width: 1300px) {
+  .btn_left,
+  .btn_right {
+    height: 0;
+    width: 0;
+    opacity: 0;
+    padding: 0;
+  }
 }
