+++ src/frontend/pages/landing/Header.css
@@ -44,6 +44,9 @@
   margin: 1rem;
   padding: 1rem 3rem;
   font-size: 1rem;
+  color: var(--white);
+  background-color: var(--tag);
+  border: 2px solid var(--tag);
   border-radius: 10px;
 }
